+++ scripts/database/schema.postgresql.sql
@@ -66,7 +66,7 @@ CREATE TABLE valid (
 CREATE TABLE users (
     id serial NOT NULL,
     login VARCHAR (200) NOT NULL,
-    pw VARCHAR (128) NOT NULL,
+    pw VARCHAR (255) NOT NULL,
     title VARCHAR (50) NULL,
     first_name VARCHAR (100) NOT NULL,
     last_name VARCHAR (100) NOT NULL,
@@ -1735,7 +1735,7 @@ CREATE TABLE customer_user (
     login VARCHAR (200) NOT NULL,
     email VARCHAR (150) NOT NULL,
     customer_id VARCHAR (150) NOT NULL,
-    pw VARCHAR (128) NULL,
+    pw VARCHAR (255) NULL,
     title VARCHAR (50) NULL,
     first_name VARCHAR (100) NOT NULL,
     last_name VARCHAR (100) NOT NULL,
@@ -1833,7 +1833,7 @@ END$$;
 CREATE TABLE mail_account (
     id serial NOT NULL,
     login VARCHAR (200) NOT NULL,
-    pw VARCHAR (200) NOT NULL,
+    pw VARCHAR (255) NOT NULL,
     host VARCHAR (200) NOT NULL,
     account_type VARCHAR (20) NOT NULL,
     queue_id INTEGER NOT NULL,